feat: 新增 History 欄位順序與欄寬的 notifier 與持久化
1. GraphColumnOrderNotifier 的 move() 拒絕動到 graph/message 兩個鎖定槽，
   套用後重新 resolve，越界或非法拖曳退化為 no-op 而非弄壞清單
2. GraphColumnWidthNotifier 把 setWidth（只動 state）與 commitWidths（寫入
   prefs）分開，避免拖曳期間每幀寫 SharedPreferences
3. 新增 graphColumnLayoutProvider，把順序、欄寬、可見性合成單一 GraphColumnLayout，
   讓 row、ladder、拖曳區三邊讀同一份
4. graph_column.dart 移至 data/models 並取代該處零呼叫端的死碼 GraphColumn class，
   模型本身不 import 任何專案檔，維持 data 層不反向依賴 features
5. 17 個測試涵蓋兩個 notifier 與 layout provider；mutation check 確認拿掉
   commitWidths 的寫入後只有「重建後仍維持」變紅
